requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
attrs==24.2.0
az-iranian-bank-gateways==2.0.9
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
Django==5.0.4
django-ckeditor==6.7.1
django-jalali==6.2.1
django-js-asset==2.2.0
django-taggit==5.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.9
inflection==0.5.1
isodate==0.6.1
jalali_core==1.0.0
jdatetime==5.0.0
lxml==5.3.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.3.0
platformdirs==4.3.3
psycopg==3.1.18
psycopg2==2.9.9
pycryptodome==3.20.0
python-bidi==0.6.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
zeep==4.2.1
